+++ src/utils/globalVar.js
@@ -1,0 +1,2 @@
+export const msgError = 'Operation failed';
+export const invalidInputError = 'Invalid input';
